feat: skip schema update
This feature helps starting server faster and it's implemented for dev purposes.
It allows the usage fo the following configuration: db.schemaUpdate: TRUE|FALSE.
When this configuration is missing, port starts as usual. The first time when
the feature is used or first time after any db schema update, it needs to be
set to TRUE and it creates 2 txt files in db/ directory which will be later read
for all subsequent runs with setting set to FALSE.